requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.2.6
django-crispy-forms==2.1
django-form-surveys==2.0.1
numpy==1.26.1
python-decouple==3.8
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
validate-email==1.3
